+++ App/models/position.py
@@ -1,0 +1,11 @@
+from enum import Enum
+
+
+class Position(Enum):
+    PTINSTRUCTOR = "Part-Time Instructor"
+    INSTRUCTOR = "Instructor"
+    HOD = "Head of Department"
+    LECTURER = "Lecturer"
+    TA = "Teaching Assistant"
+    TUTOR = "Tutor"
+    PTTUTOR = "Part-Time Tutor"
